Excluding xerces / xml-apis libs
These are provided since java 1.4, and cause class mismatch in the
classloader (when redeploying for instance) and some weird errors
related to XMLParserConfiguration cast exceptions.

Related to cigalsace raster extraction (but does not fix entirely the
raster bug).

tested at runtime with multiple redeploys, this does not harm the
extractorapp behaviour.