extractorapp - Fuzziest check on layer name
Depending on which GetCapabilities is made onto the GeoServer endpoint,
the layer names in the WCS GetCapabilities document can be or not
prefixed.

this patch proposes to be fuzzier on the layer name check, removing the
prefix.

Related Github issue:
https://github.com/georchestra/georchestra/issues/975

Tests: unit tests, not tested at runtime.

Bug: I consider checking the layer existence using regex onto XML
documents as a bug, it would be relevant imho to rewrite this part (not
to mention that this is kind of duplicated with the check present in the
WfsExtractor). The main reason it has been made this way is that some
implementations returned badly-formatted XML documents.